Implement tls-prefer-server-ciphers option to prefer server's ciphers
Prefer server ciphers over client ciphers during the TLS handshake. Part
of the ClientHello message client sends supported cipher suites in order
of its preference. If tls-prefer-server-ciphers is when server choosing
a cipher, use the server's preferences instead of the client preferences.

The functionality is the same as the folowings:

* ssl_prefer_server_ciphers in NGINX
* SSLHonorCipherOrder in Apache
* ...